Verificación antes de editar y eliminar

+++ application/models/Caracteristicas_model.php
@@ -15,6 +15,11 @@ class Caracteristicas_model extends CI_Model {
     }
     function editar($id = NULL, $atributo = NULL, $valor = NULL){
         if($id != NULL AND $atributo != NULL AND $valor != NULL){
+            if($atributo =="nombre"){
+                $this->db->where('nombre', $valor);
+                if($this->db->count_all_results()>0){
+                    return array('res'=>'bad','msj'=>'Error en la edición. Ya existe una característica con ese nombre.'); }
+            }
             $this->db->trans_start();
             $object = array($atributo => $valor);
             $this->db->where('id', $id);
@@ -44,6 +49,9 @@ class Caracteristicas_model extends CI_Model {
     function eliminar($id = null){
         if($id == null){
             return array('res'=>'bad','msj'=>'Error en la inserción.'); }
+        $this->db->where('idCaracteristica', $id);
+        if($this->db->count_all_results()>0){
+            return array('res'=>'bad','msj'=>'Error en la inserción. Hay productos asociados a esta característica.'); }
         $this->db->where('id', $id);
         $this->db->delete('caracteristicas');
         return array('res'=>'ok');
